`HoodieTableSource` does not implement SupportsPartitionPushDown anymore
