+++ hudi-flink-datasource/hudi-flink/src/test/java/org/apache/hudi/table/ITTestHoodieDataSource.java
@@ -46,6 +46,8 @@ import org.apache.flink.table.api.internal.TableEnvironmentImpl;
 import org.apache.flink.table.catalog.ObjectPath;
 import org.apache.flink.table.catalog.exceptions.TableNotExistException;
 import org.apache.flink.table.data.RowData;
+import org.apache.flink.table.data.StringData;
+import org.apache.flink.table.data.TimestampData;
 import org.apache.flink.types.Row;
 import org.apache.flink.util.CollectionUtil;
 import org.junit.jupiter.api.BeforeEach;
@@ -60,6 +62,7 @@ import org.junit.jupiter.params.provider.ValueSource;
 
 import java.io.File;
 import java.time.ZoneId;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashMap;
@@ -76,6 +79,7 @@ import static org.apache.hudi.utils.TestConfigurations.sql;
 import static org.apache.hudi.utils.TestData.array;
 import static org.apache.hudi.utils.TestData.assertRowsEqualsUnordered;
 import static org.apache.hudi.utils.TestData.assertRowsEquals;
+import static org.apache.hudi.utils.TestData.insertRow;
 import static org.apache.hudi.utils.TestData.map;
 import static org.apache.hudi.utils.TestData.row;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -86,6 +90,17 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
  */
 @ExtendWith(FlinkMiniCluster.class)
 public class ITTestHoodieDataSource {
+
+  private static List<RowData> DATA_SET_NEW_PARTITIONS = Arrays.asList(
+      insertRow(
+          StringData.fromString("id9"), StringData.fromString("LiLi"), 24,
+          TimestampData.fromEpochMillis(9), StringData.fromString("par5")),
+      insertRow(StringData.fromString("id11"), StringData.fromString("Baobao"), 34,
+          TimestampData.fromEpochMillis(10), StringData.fromString("par7")),
+      insertRow(StringData.fromString("id12"), StringData.fromString("LinLin"), 34,
+          TimestampData.fromEpochMillis(10), StringData.fromString("part8"))
+  );
+
   private TableEnvironment streamTableEnv;
   private TableEnvironment batchTableEnv;
 
@@ -1764,6 +1779,91 @@ public class ITTestHoodieDataSource {
   }
 
   // -------------------------------------------------------------------------
+  //  Test continuous partition prune
+  // -------------------------------------------------------------------------
+
+  @ParameterizedTest
+  @MethodSource("tableTypeAndPartitioningParams")
+  void testOr(
+      HoodieTableType tableType, boolean hiveStylePartitioning) throws Exception {
+    String condition = "`partition` = 'par5' or `partition` = 'par6'";
+    List<RowData> result = Arrays.asList(
+        insertRow(StringData.fromString("id9"), StringData.fromString("LiLi"), 24,
+            TimestampData.fromEpochMillis(9), StringData.fromString("par5")));
+    testContinuousPartitionPrune(tableType, hiveStylePartitioning, condition, result);
+  }
+
+  @ParameterizedTest
+  @MethodSource("tableTypeAndPartitioningParams")
+  void testAnd(
+      HoodieTableType tableType, boolean hiveStylePartitioning) throws Exception {
+    String condition = "`partition` >= 'par5' and `partition` <= 'par6'";
+    List<RowData> result = Arrays.asList(
+        insertRow(StringData.fromString("id9"), StringData.fromString("LiLi"), 24,
+            TimestampData.fromEpochMillis(9), StringData.fromString("par5")));
+    testContinuousPartitionPrune(tableType, hiveStylePartitioning, condition, result);
+  }
+
+  @ParameterizedTest
+  @MethodSource("tableTypeAndPartitioningParams")
+  void testNot(
+      HoodieTableType tableType, boolean hiveStylePartitioning) throws Exception {
+    String condition = "`partition` <> 'par7'";
+    List<RowData> result = Arrays.asList(
+        insertRow(StringData.fromString("id1"), StringData.fromString("Danny"), 23,
+            TimestampData.fromEpochMillis(1), StringData.fromString("par1")),
+        insertRow(StringData.fromString("id2"), StringData.fromString("Stephen"), 33,
+            TimestampData.fromEpochMillis(2), StringData.fromString("par1")),
+        insertRow(StringData.fromString("id3"), StringData.fromString("Julian"), 53,
+            TimestampData.fromEpochMillis(3), StringData.fromString("par2")),
+        insertRow(StringData.fromString("id4"), StringData.fromString("Fabian"), 31,
+            TimestampData.fromEpochMillis(4), StringData.fromString("par2")),
+        insertRow(StringData.fromString("id5"), StringData.fromString("Sophia"), 18,
+            TimestampData.fromEpochMillis(5), StringData.fromString("par3")),
+        insertRow(StringData.fromString("id6"), StringData.fromString("Emma"), 20,
+            TimestampData.fromEpochMillis(6), StringData.fromString("par3")),
+        insertRow(StringData.fromString("id7"), StringData.fromString("Bob"), 44,
+            TimestampData.fromEpochMillis(7), StringData.fromString("par4")),
+        insertRow(StringData.fromString("id8"), StringData.fromString("Han"), 56,
+            TimestampData.fromEpochMillis(8), StringData.fromString("par4")),
+        insertRow(StringData.fromString("id9"), StringData.fromString("LiLi"), 24,
+            TimestampData.fromEpochMillis(9), StringData.fromString("par5")),
+        insertRow(StringData.fromString("id12"), StringData.fromString("LinLin"), 34,
+            TimestampData.fromEpochMillis(10), StringData.fromString("part8"))
+    );
+    testContinuousPartitionPrune(tableType, hiveStylePartitioning, condition, result);
+  }
+
+  @ParameterizedTest
+  @MethodSource("tableTypeAndPartitioningParams")
+  void testNotIn(
+      HoodieTableType tableType, boolean hiveStylePartitioning) throws Exception {
+    String condition = "`partition` not in ('par1', 'par2', 'par3')";
+    List<RowData> result = Arrays.asList(
+        insertRow(StringData.fromString("id7"), StringData.fromString("Bob"), 44,
+            TimestampData.fromEpochMillis(7), StringData.fromString("par4")),
+        insertRow(StringData.fromString("id8"), StringData.fromString("Han"), 56,
+            TimestampData.fromEpochMillis(8), StringData.fromString("par4")),
+        insertRow(StringData.fromString("id9"), StringData.fromString("LiLi"), 24,
+            TimestampData.fromEpochMillis(9), StringData.fromString("par5")),
+        insertRow(StringData.fromString("id11"), StringData.fromString("Baobao"), 34,
+            TimestampData.fromEpochMillis(10), StringData.fromString("par7")),
+        insertRow(StringData.fromString("id12"), StringData.fromString("LinLin"), 34,
+            TimestampData.fromEpochMillis(10), StringData.fromString("part8"))
+    );
+    testContinuousPartitionPrune(tableType, hiveStylePartitioning, condition, result);
+  }
+
+  @ParameterizedTest
+  @MethodSource("tableTypeAndPartitioningParams")
+  void testContainNonSimpleCall(
+      HoodieTableType tableType, boolean hiveStylePartitioning) throws Exception {
+    String condition = "`partition` like 'part%' and `partition` <> 'part8' ";
+    List<RowData> result = new ArrayList<>();
+    testContinuousPartitionPrune(tableType, hiveStylePartitioning, condition, result);
+  }
+
+  // -------------------------------------------------------------------------
   //  Utilities
   // -------------------------------------------------------------------------
   private enum ExecMode {
@@ -1884,6 +1984,68 @@ public class ITTestHoodieDataSource {
     tEnv.executeSql("DROP TABLE IF EXISTS sink");
     tEnv.executeSql(sinkDDL);
     TableResult tableResult = tEnv.executeSql("insert into sink " + select);
+    // wait for the timeout then cancels the job
+    TimeUnit.SECONDS.sleep(timeout);
+    tableResult.getJobClient().ifPresent(JobClient::cancel);
+    tEnv.executeSql("DROP TABLE IF EXISTS sink");
+    return CollectSinkTableFactory.RESULT.values().stream()
+        .flatMap(Collection::stream)
+        .collect(Collectors.toList());
+  }
+
+  private void testContinuousPartitionPrune(
+      HoodieTableType tableType,
+      boolean hiveStylePartitioning,
+      String filterCondition,
+      List<RowData> results
+  ) throws Exception {
+    Configuration conf = TestConfigurations.getDefaultConf(tempFile.getAbsolutePath());
+    conf.setString(FlinkOptions.TABLE_NAME, "t1");
+    conf.setString(FlinkOptions.TABLE_TYPE, tableType.name());
+    conf.setBoolean(FlinkOptions.HIVE_STYLE_PARTITIONING, hiveStylePartitioning);
+
+    // write one commit
+    TestData.writeData(TestData.DATA_SET_INSERT, conf);
+
+    String hoodieTableDDL = sql("t1")
+        .option(FlinkOptions.PATH, tempFile.getAbsolutePath())
+        .option(FlinkOptions.TABLE_TYPE, tableType)
+        .option(FlinkOptions.READ_AS_STREAMING, true)
+        .option(FlinkOptions.READ_STREAMING_CHECK_INTERVAL, 2)
+        .option(FlinkOptions.HIVE_STYLE_PARTITIONING, hiveStylePartitioning)
+        .end();
+    streamTableEnv.executeSql(hoodieTableDDL);
+
+    String sinkDDL = "create table sink(\n"
+        + "  uuid varchar(20),\n"
+        + "  name varchar(20),\n"
+        + "  age int,\n"
+        + "  ts timestamp,\n"
+        + "  part varchar(20)"
+        + ") with (\n"
+        + "  'connector' = '" + CollectSinkTableFactory.FACTORY_ID + "'"
+        + ")";
+    TableResult tableResult = submitSelectSql(
+        streamTableEnv,
+        "select uuid, name, age, ts, `partition` as part from t1 where " + filterCondition,
+        sinkDDL);
+
+    // write second commit
+    TestData.writeData(DATA_SET_NEW_PARTITIONS, conf);
+    // stop the streaming query and fetch the result
+    List<Row> result = stopAndFetchData(streamTableEnv, tableResult, 10);
+    assertRowsEquals(result, results);
+  }
+
+  private TableResult submitSelectSql(TableEnvironment tEnv, String select, String sinkDDL) {
+    tEnv.executeSql("DROP TABLE IF EXISTS sink");
+    tEnv.executeSql(sinkDDL);
+    TableResult tableResult = tEnv.executeSql("insert into sink " + select);
+    return tableResult;
+  }
+
+  private List<Row> stopAndFetchData(TableEnvironment tEnv, TableResult tableResult, long timeout)
+      throws InterruptedException {
     // wait for the timeout then cancels the job
     TimeUnit.SECONDS.sleep(timeout);
     tableResult.getJobClient().ifPresent(JobClient::cancel);
