+++ hudi-flink-datasource/hudi-flink/src/test/java/org/apache/hudi/table/format/TestInputFormat.java
@@ -29,7 +29,7 @@ import org.apache.hudi.common.table.timeline.HoodieTimeline;
 import org.apache.hudi.configuration.FlinkOptions;
 import org.apache.hudi.configuration.HadoopConfigurations;
 import org.apache.hudi.source.IncrementalInputSplits;
-import org.apache.hudi.source.prune.StaticPartitionPruner;
+import org.apache.hudi.source.prune.PartitionPruner;
 import org.apache.hudi.table.HoodieTableSource;
 import org.apache.hudi.table.format.cdc.CdcInputFormat;
 import org.apache.hudi.table.format.cow.CopyOnWriteInputFormat;
@@ -45,7 +45,12 @@ import org.apache.hudi.utils.TestUtils;
 import org.apache.flink.api.common.io.InputFormat;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.core.io.InputSplit;
+import org.apache.flink.table.api.DataTypes;
 import org.apache.flink.table.data.RowData;
+import org.apache.flink.table.expressions.CallExpression;
+import org.apache.flink.table.expressions.FieldReferenceExpression;
+import org.apache.flink.table.expressions.ValueLiteralExpression;
+import org.apache.flink.table.functions.BuiltInFunctionDefinitions;
 import org.apache.hadoop.fs.Path;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
@@ -341,7 +346,7 @@ public class TestInputFormat {
         .rowType(TestConfigurations.ROW_TYPE)
         .conf(conf)
         .path(FilePathUtils.toFlinkPath(metaClient.getBasePathV2()))
-        .partitionPruner(new StaticPartitionPruner(Collections.emptySet()))
+        .partitionPruner(new PartitionPruner(Collections.emptySet()))
         .skipCompaction(false)
         .build();
 
@@ -383,7 +388,7 @@ public class TestInputFormat {
         .rowType(TestConfigurations.ROW_TYPE)
         .conf(conf)
         .path(FilePathUtils.toFlinkPath(metaClient.getBasePathV2()))
-        .partitionPruner(new StaticPartitionPruner(new HashSet<>(Arrays.asList("par1", "par2", "par3", "par4"))))
+        .partitionPruner(new PartitionPruner(new HashSet<>(Arrays.asList("par1", "par2", "par3", "par4"))))
         .skipCompaction(false)
         .build();
 
@@ -422,7 +427,7 @@ public class TestInputFormat {
         .rowType(TestConfigurations.ROW_TYPE)
         .conf(conf)
         .path(FilePathUtils.toFlinkPath(metaClient.getBasePathV2()))
-        .partitionPruner(new StaticPartitionPruner(new HashSet<>(Arrays.asList("par1", "par2", "par3", "par4"))))
+        .partitionPruner(new PartitionPruner(new HashSet<>(Arrays.asList("par1", "par2", "par3", "par4"))))
         .skipCompaction(true)
         .build();
 
@@ -489,7 +494,7 @@ public class TestInputFormat {
         .rowType(TestConfigurations.ROW_TYPE)
         .conf(conf)
         .path(FilePathUtils.toFlinkPath(metaClient.getBasePathV2()))
-        .partitionPruner(new StaticPartitionPruner(new HashSet<>(Arrays.asList("par1", "par2", "par3", "par4"))))
+        .partitionPruner(new PartitionPruner(new HashSet<>(Arrays.asList("par1", "par2", "par3", "par4"))))
         .skipClustering(true)
         .build();
 
@@ -553,7 +558,7 @@ public class TestInputFormat {
         .rowType(TestConfigurations.ROW_TYPE)
         .conf(conf)
         .path(FilePathUtils.toFlinkPath(metaClient.getBasePathV2()))
-        .partitionPruner(new StaticPartitionPruner(new HashSet<>(Arrays.asList("par1", "par2", "par3", "par4"))))
+        .partitionPruner(new PartitionPruner(new HashSet<>(Arrays.asList("par1", "par2", "par3", "par4"))))
         .build();
 
     // default read the latest commit
@@ -587,10 +592,14 @@ public class TestInputFormat {
 
     TestData.writeData(TestData.DATA_SET_INSERT, conf);
 
-    Map<String, String> prunedPartitions = new HashMap<>();
-    prunedPartitions.put("partition", "par1");
     // prune to only be with partition 'par1'
-    tableSource.applyPartitions(Collections.singletonList(prunedPartitions));
+    FieldReferenceExpression partRef = new FieldReferenceExpression("partition", DataTypes.STRING(), 4, 4);
+    ValueLiteralExpression partLiteral = new ValueLiteralExpression("par1", DataTypes.STRING().notNull());
+    CallExpression partFilter = new CallExpression(
+        BuiltInFunctionDefinitions.EQUALS,
+        Arrays.asList(partRef, partLiteral),
+        DataTypes.BOOLEAN());
+    tableSource.applyFilters(Arrays.asList(partFilter));
     InputFormat<RowData, ?> inputFormat = tableSource.getInputFormat();
 
     List<RowData> result = readData(inputFormat);
